+++ libreadtags/NEWS.md
@@ -1,3 +1,11 @@
+# Version ???
+
+- read input fields, values at the second column in a tag file, with
+  unescaping if !_TAG_OUTPUT_MODE is "u-ctags" and
+  !_TAG_OUTPUT_FILESEP is "slash" in the tag file.
+
+- LT_VERSION ?:?:?
+
 # Version 0.2.1
 
 - use "m" mode flag of fopen only when compiling with glibc 2.3 or higher.